Cat Tanks and Submission 2 test
All cat tanks now in game. I see that i need to update the format of the guns, and will adjust them appropriately. They are in the Characters folder in the Textures.
The Submission 2 test is up, in the file with the GDD and One Sheet. Please read and send feedback :)